Improving UI Design

+++ app/src/main/java/com/example/uas_pam/ui/detail/DetailScreen.kt
@@ -1,19 +1,26 @@
 package com.example.uas_pam.ui.detail
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExtendedFloatingActionButton
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -29,14 +36,22 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.uas_pam.PenyediaViewModel
+import com.example.uas_pam.R
 import com.example.uas_pam.navigation.DestinasiNavigasi
 import com.example.uas_pam.ui.AllData
 import com.example.uas_pam.ui.DetailUIState
 import com.example.uas_pam.ui.ImtTopAppBar
+import com.example.uas_pam.ui.theme.UAS_PAMTheme
 import com.example.uas_pam.ui.toAllData
 import kotlinx.coroutines.launch
 
@@ -58,6 +73,7 @@ fun DetailScreen(
 ) {
     val uiState = viewModel.uiState.collectAsState()
     val coroutineScope = rememberCoroutineScope()
+    var deleteConfirmationRequired by rememberSaveable { mutableStateOf(false) }
     Scaffold(
         topBar = {
             ImtTopAppBar(
@@ -66,31 +82,50 @@ fun DetailScreen(
                 navigateUp = navigateBack
             )
         }, floatingActionButton = {
-            FloatingActionButton(
-                onClick = { navigateToEditItem(uiState.value.allDataUi.id) },
-                shape = MaterialTheme.shapes.medium,
-                modifier = Modifier.padding(18.dp)
-
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Edit,
-                    contentDescription = "",
+            Row (
+                horizontalArrangement = Arrangement.SpaceBetween,
+                modifier = Modifier.padding(4.dp),
+            ){
+                ExtendedFloatingActionButton(
+                    onClick = {  deleteConfirmationRequired = true },
+                    icon = { Icon(Icons.Filled.Delete, "") },
+                    text = { Text(text = "Delete",fontFamily = FontFamily.Serif ) },
+                    containerColor = Color.LightGray
+                )
+                Spacer(modifier = Modifier.width(16.dp))
+                ExtendedFloatingActionButton(
+                    onClick = { navigateToEditItem(uiState.value.allDataUi.id) },
+                    icon = { Icon(Icons.Filled.Edit, "") },
+                    text = { Text(text = "Edit", fontFamily = FontFamily.Serif) },
+                    containerColor = Color.LightGray
+                )
+            }
+            if (deleteConfirmationRequired) {
+                DeleteConfirmationDialog(
+                    onDeleteConfirm = {
+                        deleteConfirmationRequired = false
+                        coroutineScope.launch {
+                            viewModel.deleteKontak()
+                            navigateBack()
+                        }
+                    },
+                    onDeleteCancel = { deleteConfirmationRequired = false },
+                    modifier = Modifier.padding(12.dp)
                 )
             }
         }, modifier = modifier
     ) { innerPadding ->
+        Box(modifier = Modifier.fillMaxSize()) {
+            Image(
+                painter = painterResource(id = R.drawable.back),
+                contentDescription = "",
+                contentScale = ContentScale.FillBounds,
+                modifier = Modifier.matchParentSize()
+            )
+        }
         ItemDetailsBody(
             detailUIState = uiState.value,
-            onDelete = {
-                // Note: If the user rotates the screen very fast, the operation may get cancelled
-                // and the item may not be deleted from the Database. This is because when config
-                // change occurs, the Activity will be recreated and the rememberCoroutineScope will
-                // be cancelled - since the scope is bound to composition.
-                coroutineScope.launch {
-                    viewModel.deleteKontak()
-                    navigateBack()
-                }
-            },
+
             modifier = Modifier
                 .padding(innerPadding)
                 .verticalScroll(rememberScrollState()),
@@ -101,35 +136,15 @@ fun DetailScreen(
 @Composable
 private fun ItemDetailsBody(
     detailUIState: DetailUIState,
-    onDelete: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Column(
         modifier = modifier.padding(12.dp),
         verticalArrangement = Arrangement.spacedBy(12.dp)
     ) {
-        var deleteConfirmationRequired by rememberSaveable { mutableStateOf(false) }
         ItemDetails(
             allData  = detailUIState.allDataUi.toAllData(), modifier = Modifier.fillMaxWidth()
         )
-
-        OutlinedButton(
-            onClick = { deleteConfirmationRequired = true },
-            shape = MaterialTheme.shapes.small,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text("Delete")
-        }
-        if (deleteConfirmationRequired) {
-            DeleteConfirmationDialog(
-                onDeleteConfirm = {
-                    deleteConfirmationRequired = false
-                    onDelete()
-                },
-                onDeleteCancel = { deleteConfirmationRequired = false },
-                modifier = Modifier.padding(12.dp)
-            )
-        }
     }
 }
 
@@ -138,10 +153,7 @@ fun ItemDetails(
     allData: AllData, modifier: Modifier = Modifier
 ) {
     Card(
-        modifier = modifier, colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.primaryContainer,
-            contentColor = MaterialTheme.colorScheme.onPrimaryContainer
-        )
+        modifier = modifier, colors = CardDefaults.cardColors(Color.LightGray)
     ) {
         Column(
             modifier = Modifier
@@ -200,9 +212,9 @@ private fun ItemDetailsRow(
     labelResID: String, itemDetail: String, modifier: Modifier = Modifier
 ) {
     Row(modifier = modifier) {
-        Text(text = labelResID, fontWeight = FontWeight.Bold)
+        Text(text = labelResID, fontWeight = FontWeight.Bold,fontFamily = FontFamily.Serif,)
         Spacer(modifier = Modifier.weight(1f))
-        Text(text = itemDetail, fontWeight = FontWeight.Bold)
+        Text(text = itemDetail, fontStyle = FontStyle.Italic,fontFamily = FontFamily.Serif,)
     }
 }
 
@@ -211,17 +223,27 @@ private fun DeleteConfirmationDialog(
     onDeleteConfirm: () -> Unit, onDeleteCancel: () -> Unit, modifier: Modifier = Modifier
 ) {
     AlertDialog(onDismissRequest = { /* Do nothing */ },
-        title = { Text("Are you sure") },
-        text = { Text("Delete") },
+        title = { Text("Are you sure?",fontFamily = FontFamily.Serif) },
+        text = { Text("Delete", fontFamily = FontFamily.Serif) },
         modifier = modifier,
         dismissButton = {
             TextButton(onClick = onDeleteCancel) {
-                Text(text = "No")
+                Text(text = "No", fontFamily = FontFamily.Serif)
             }
         },
         confirmButton = {
             TextButton(onClick = onDeleteConfirm) {
-                Text(text = "Yes")
+                Text(text = "Yes", fontFamily = FontFamily.Serif)
             }
-        })
+        },
+        containerColor = Color.LightGray
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun DetailScreenPreview() {
+    UAS_PAMTheme {
+        DetailScreen(navigateToEditItem = {} , navigateBack = { /*TODO*/ })
+    }
 }
